+++ tests/gateway/test_busy_steer_origin.py
@@ -5,7 +5,7 @@ import json
 import pytest
 
 from gateway.config import GatewayConfig, Platform
-from gateway.platforms.base import MessageEvent
+from gateway.platforms.event import MessageEvent
 from gateway.run import GatewayRunner
 from gateway.session import SessionSource
